Separate components, container, pages, modules to each folder
