+++ src/containers/RestaurantContainer.jsx
@@ -2,17 +2,17 @@ import React, { useEffect } from 'react';
 
 import { useDispatch, useSelector } from 'react-redux';
 
-import RestaurantDetail from './RestaurantDetail';
-import ReviewForm from './ReviewForm';
-import Reviews from './Reviews';
+import RestaurantDetail from '../components/RestaurantDetail';
+import ReviewForm from '../components/ReviewForm';
+import Reviews from '../components/Reviews';
 
 import {
   loadRestaurant,
   changeReviewField,
   sendReview,
-} from './actions';
+} from '../modules/actions';
 
-import { get } from './utils';
+import { get } from '../modules/utils';
 
 export default function RestaurantContainer({ restaurantId }) {
   const dispatch = useDispatch();
