+++ src/index.jsx
@@ -9,7 +9,7 @@ import { Provider } from 'react-redux';
 
 import App from './App';
 
-import store from './store';
+import store from './modules/store';
 
 ReactDOM.render(
   (
