+++ src/modules/actions.test.js
@@ -21,7 +21,7 @@ import {
 const middlewares = [thunk];
 const mockStore = configureStore(middlewares);
 
-jest.mock('./services/api');
+jest.mock('../services/api');
 
 describe('actions', () => {
   let store;
