+++ src/pages/RestaurantsPage.jsx
@@ -4,13 +4,13 @@ import { useHistory } from 'react-router-dom';
 
 import { useDispatch } from 'react-redux';
 
-import RegionsContainer from './RegionsContainer';
-import CategoriesContainer from './CategoriesContainer';
-import RestaurantsContainer from './RestaurantsContainer';
+import RegionsContainer from '../containers/RegionsContainer';
+import CategoriesContainer from '../containers/CategoriesContainer';
+import RestaurantsContainer from '../containers/RestaurantsContainer';
 
 import {
   loadInitialData,
-} from './actions';
+} from '../modules/actions';
 
 export default function RestaurantsPage() {
   const history = useHistory();
